refactor: rename package structure for Notes-related classes to improve organization
